split(store/sqlite): break store_authz_listing.go into three topic files
store_authz_listing.go was 538 lines, 38 lines over the T65 500-line
limit. Split by principal side to mirror the existing test-file split:

- store_authz_user_listing.go  (137 lines) — UserAuthzResourcesList,
  shared SQL/helpers (userAuthzResourcesBaseSQL, inPlaceholders,
  buildUserAuthzMaskQueryAndArgs, userEffectivePermissionArgs)
- store_authz_group_listing.go (116 lines) — GroupAuthzResourcesList
  and groupAuthzResourcesBaseSQL
- store_authz_resource_listing.go (305 lines) — ResourceAuthzGroupsList,
  ResourceAuthzUsersList and their SQL constants plus scanUserMasks

All three new files have alphabetically ordered imports.
